requirements.txt
django>=5.1,<5.2
psycopg2-binary>=2.9
gunicorn>=22.0
django-environ>=0.11
whitenoise>=6.7
pillow>=10.0
django-q2>=1.6
croniter>=1.4
sendgrid>=6.11
